Switch to Windows-only

+++ examples/00_clip/src/example_clip/app_example_clip.cpp
@@ -84,7 +84,7 @@ app_example_clip::app_example_clip(const unsigned int width,
 
 void app_example_clip::update(const float dt_s)
 {
-  bgfx::setViewRect(view_id, 0, 0, get_width(), get_height());
+  bgfx::setViewRect(view_id, 0, 0, gsl::narrow<uint16_t>(get_width()), gsl::narrow<uint16_t>(get_height()));
   bgfx::setViewClear(view_id, BGFX_CLEAR_COLOR | BGFX_CLEAR_DEPTH, view_clear_color);
 
   _scene.update(dt_s);
