+++ examples/03_state_machine_simple/src/example_state_machine_simple/app_example_state_machine_simple.cpp
@@ -193,7 +193,8 @@ app_example_state_machine_simple::app_example_state_machine_simple(const unsigne
 
 void app_example_state_machine_simple::update(const float dt_s)
 {
-  bgfx::setViewRect(view_id, 0, 0, get_width(), get_height());
+  bgfx::setViewRect(view_id, 0, 0, gsl::narrow<uint16_t>(get_width()),
+                    gsl::narrow<uint16_t>(get_height()));
   bgfx::setViewClear(view_id, BGFX_CLEAR_COLOR | BGFX_CLEAR_DEPTH, view_clear_color);
 
   ImGui::SetNextWindowSize(ImVec2(350.0F, 0.0F));
